Fix Weather Data Links
The LPO Weather Data website has changed its structure such that the
links in this project that attempt to download the data no longer
function. Lynda.com have provided a github hosted copy of the data as
a replacement. This change moves references to the old site to the new
copy so that exercises can be completed.